Extract unified BenchJdbcJob from BenchJdbcAvroJob and BenchJdbcParquetJob
The two benchmark classes were ~98% identical, differing only in which
job class they instantiate. Introduce BenchJdbcJob parameterized by a
JobRunner functional interface and a job name string.

BenchJdbcAvroJob and BenchJdbcParquetJob are retained as thin wrappers
that supply the format-specific job factory, preserving their main()
entry points and test compatibility.
