+++ dbeam-parquet/src/main/java/com/spotify/dbeam/parquet/BenchJdbcParquetJob.java
@@ -20,114 +20,20 @@
 
 package com.spotify.dbeam.parquet;
 
-import static com.google.common.collect.Lists.newArrayList;
-
-import com.google.common.collect.ImmutableMap;
-import com.google.common.math.Stats;
-import com.spotify.dbeam.beam.MetricsHelper;
+import com.spotify.dbeam.jobs.BenchJdbcJob;
 import com.spotify.dbeam.jobs.ExceptionHandling;
-import com.spotify.dbeam.options.OutputOptions;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collector;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
-import org.apache.beam.sdk.PipelineResult;
-import org.apache.beam.sdk.options.Default;
-import org.apache.beam.sdk.options.Description;
-import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 
 /** Used on e2e test, allows benchmarking with different configuration parameters. */
 public class BenchJdbcParquetJob {
 
-  public interface BenchJdbcParquetOptions extends PipelineOptions {
-    @Description("Number of benchmark executions.")
-    @Default.Integer(3)
-    int getExecutions();
-
-    void setExecutions(int value);
-  }
-
-  private final PipelineOptions pipelineOptions;
-  private List<Map<String, Long>> metrics = newArrayList();
-
-  public BenchJdbcParquetJob(final PipelineOptions pipelineOptions) {
-    this.pipelineOptions = pipelineOptions;
-  }
-
-  public static BenchJdbcParquetJob create(final String[] cmdLineArgs) {
-    PipelineOptionsFactory.register(BenchJdbcParquetOptions.class);
-    PipelineOptions options = JdbcParquetJob.buildPipelineOptions(cmdLineArgs);
-    return new BenchJdbcParquetJob(options);
-  }
-
-  public void run() throws Exception {
-    int executions = pipelineOptions.as(BenchJdbcParquetOptions.class).getExecutions();
-    for (int i = 0; i < executions; i++) {
-      String output =
-          String.format("%s/run_%d", pipelineOptions.as(OutputOptions.class).getOutput(), i);
-      final PipelineResult pipelineResult =
-          JdbcParquetJob.create(pipelineOptions, output).runExport();
-      this.metrics.add(MetricsHelper.getMetrics(pipelineResult));
-    }
-    System.out.println("Summary for BenchJdbcParquetJob");
-    System.out.println(pipelineOptions.toString());
-    System.out.println(tsvMetrics());
-  }
-
-  private String tsvMetrics() {
-    final List<String> columns =
-        newArrayList(
-            "recordCount", "writeElapsedMs", "msPerMillionRows", "bytesWritten", "KbWritePerSec");
-    final Collector<CharSequence, ?, String> tabJoining = Collectors.joining("\t");
-    final Stream<String> lines =
-        IntStream.range(0, this.metrics.size())
-            .mapToObj(
-                i ->
-                    String.format(
-                        "run_%02d  \t%s",
-                        i,
-                        columns.stream()
-                            .map(
-                                c ->
-                                    String.format(
-                                        "% 10d",
-                                        Optional.of(this.metrics.get(i).get(c)).orElse(0L)))
-                            .collect(tabJoining)));
-    final List<Stats> stats =
-        columns.stream()
-            .map(
-                c ->
-                    Stats.of(
-                        (Iterable<Long>)
-                            this.metrics.stream().map(m -> Optional.of(m.get(c)).orElse(0L))
-                                ::iterator))
-            .collect(Collectors.toList());
-    final Map<String, Function<Stats, Double>> relevantStats =
-        ImmutableMap.of(
-            "max     ", Stats::max,
-            "mean    ", Stats::mean,
-            "min     ", Stats::min,
-            "stddev  ", Stats::populationStandardDeviation);
-    final Stream<String> statsSummary =
-        relevantStats.entrySet().stream()
-            .map(
-                e ->
-                    String.format(
-                        "%s\t%s",
-                        e.getKey(),
-                        stats.stream()
-                            .map(e.getValue())
-                            .map(v -> String.format("% 10.1f", v))
-                            .collect(tabJoining)));
-    return String.format(
-        "name    \t%12s\n%s",
-        String.join("\t", columns),
-        Stream.concat(lines, statsSummary).collect(Collectors.joining("\n")));
+  public static BenchJdbcJob create(final String[] cmdLineArgs) {
+    PipelineOptionsFactory.register(BenchJdbcJob.BenchJdbcOptions.class);
+    return BenchJdbcJob.create(
+        "BenchJdbcParquetJob",
+        cmdLineArgs,
+        JdbcParquetJob::buildPipelineOptions,
+        (opts, output) -> JdbcParquetJob.create(opts, output).runExport());
   }
 
   public static void main(String[] cmdLineArgs) {
